+++ TableViewKit/Protocols/TableItem.swift
@@ -1,8 +1,11 @@
 import Foundation
 
+@available(*, deprecated, renamed: "TableItem")
+public typealias Item = TableItem
+
 /// A type that represent an item to be displayed
 /// defining the `drawer` and the `height`
-public protocol Item: class, AnyEquatable {
+public protocol TableItem: class, AnyEquatable {
 
     /// The `drawer` of the item
     static var drawer: AnyCellDrawer { get }
@@ -11,7 +14,7 @@ public protocol Item: class, AnyEquatable {
     var height: Height? { get }
 }
 
-public extension Item where Self: Equatable {
+public extension TableItem where Self: Equatable {
     func equals(_ other: Any?) -> Bool {
         if let other = other as? Self {
             return other == self
@@ -23,7 +26,7 @@ public extension Item where Self: Equatable {
 // swiftlint:disable:next identifier_name
 private var ItemTableViewManagerKey: UInt8 = 0
 
-extension Item {
+extension TableItem {
 
     public internal(set) var manager: TableViewManager? {
         get {
@@ -38,7 +41,7 @@ extension Item {
     }
 }
 
-extension Item {
+extension TableItem {
 
     public func equals(_ other: Any?) -> Bool {
         if let other = other as AnyObject? {
@@ -55,7 +58,7 @@ extension Item {
     /// Returns the `section` of the `item`
     ///
     /// - returns: The `section` of the `item` or `nil` if not present
-    public var section: Section? {
+    public var section: TableSection? {
         guard let indexPath = indexPath else { return nil }
         return manager?.sections[indexPath.section]
     }
@@ -74,22 +77,33 @@ extension Item {
         return nil
     }
 
+    /// Returns the `cell` of the `item`, if visibile
+    var cell: UITableViewCell? {
+        guard let indexPath = indexPath else { return nil }
+        return manager?.tableView.cellForRow(at: indexPath)
+    }
+
     /// Reload the `item` with an `animation`
     ///
     /// - parameter animation: A constant that indicates how the reloading is to be animated
-    ///
-    /// - returns: The `section` of the `item` or `nil` if not present
     public func reload(with animation: UITableViewRowAnimation = .automatic) {
         guard let indexPath = indexPath else { return }
         let section = manager?.sections[indexPath.section]
         section?.items.callback?(.updates([indexPath.row]))
     }
 
+    /// Redraw the associated `cell` of the `item` without reloading it
+    /// The `draw` method of `CellDrawer` get called if the `cell` is visible
+    public func redraw() {
+        guard let cell = cell else { return }
+        Self.drawer.draw(cell, self)
+    }
+
 }
 
-public extension Collection where Self.Iterator.Element == Item {
+public extension Collection where Self.Iterator.Element == TableItem {
     /// Return the index of the `element` inside a collection of items
-    func index(of element: Item) -> Self.Index? {
+    func index(of element: TableItem) -> Self.Index? {
         return index(where: { $0 === element })
     }
 }