+++ build.sh
@@ -4,6 +4,6 @@ set -o pipefail &&
 time xcodebuild clean test \
     -workspace TableViewKit.xcworkspace \
     -scheme TableViewKit \
-    -sdk iphonesimulator10.3 \
-    -destination 'platform=iOS Simulator,name=iPhone 6s,OS=10.0' \
+    -sdk iphonesimulator11.0 \
+    -destination 'platform=iOS Simulator,name=iPhone 7,OS=11.0' \
 | xcpretty